+++ Source/GraphicPipelineOpenGLv1_0/GraphicPipelineOpenGLv1_0.cs
@@ -1708,7 +1708,7 @@ namespace GraphicPipelineOpenGLv1_0
             Gl.glEnd();
         }
 
-        public void DrawSprite(int textureId, int xCount, int yCount, int xBild, int yBild, int x, int y, int width, int height, Color colorFactor)
+        public void DrawSprite(int textureId, int xCount, int yCount, int xBild, int yBild, float x, float y, float width, float height, Color colorFactor)
         {
             float xf = 1.0f / xCount, yf = 1.0f / yCount;
             UseAlphaBlendingAndDiscardTransparent(colorFactor); // SetBlendingWithAlpha();
